requirements.txt
django
django-cities-light
django-leaflet
mysqlclient
